requirements.txt
attrs==25.4.0
autocommand==2.2.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
constantly==23.10.4
cryptography==46.0.3
filelock==3.20.0
greenlet==3.2.4
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
lensfun==0.3.4
libmsym==0.2.4
lit==21.1.5.dev0
more-itertools==10.8.0
netsnmp-python==1.0a1
numpy==2.3.4
OpenTimelineIO==0.18.0
packaging==25.0
pandas==2.3.1
platformdirs==4.5.0
psutil==7.1.3
psycopg2==2.9.11
pyaaf2==1.7.1
pycparser==2.23
pygdbmi==0.11.0.0
PyGObject==3.54.5
PySide6==6.10.0
python-dateutil==2.9.0
python-dotenv==1.2.1
pytz==2025.2
QtPy==2.4.3
requests==2.32.5
sentry-sdk==2.43.0
setuptools==80.9.0
shiboken6==6.10.0
shiboken6-generator==6.10.0
six==1.17.0
smbus==1.1
spglib==2.6.1.dev0+gc633404.d20250310
SQLAlchemy==2.0.44
TBB==0.2
Twisted==24.7.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
yt-dlp==2025.10.22
ytmusicapi==1.11.1
zope.interface==8.0.1
